[feature/share]: refactor feature/share -> feature/mypage/share
