+++ app/src/main/java/com/teampophory/pophory/data/network/model/share/ShareResponse.kt
@@ -1,6 +1,6 @@
 package com.teampophory.pophory.data.network.model.share
 
-import com.teampophory.pophory.feature.share.model.PhotoItem
+import com.teampophory.pophory.feature.home.mypage.share.model.PhotoItem
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
